Added Tags and Keys
Added a fire object that is turned off if you present it with the correct key. There are also 4 keys that you can come across and it will attach it to your player so you can bring the key to the fire object to turn it off so it does not damage the player.